auth-service/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
pyjwt
bcrypt
